switch int to long

+++ src/InlineProgressBarCtrl.h
@@ -2,9 +2,9 @@
 #include <windows.h>
 #include <commctrl.h>
 
-HWND InlineProgressBarCtrl_Create(HWND hwndParent, const int nCurrentValue, const int MaxValue, const BOOL bSmooth, const int nPane);
-BOOL InlineProgressBarCtrl_SetRange(HWND hwnd, const int nLower, const int nUpper, const int nStep);
-int InlineProgressBarCtrl_SetStep(HWND hwnd, const int nStep);
+HWND InlineProgressBarCtrl_Create(HWND hwndParent, const long nCurrentValue, const long MaxValue, const BOOL bSmooth, const int nPane);
+BOOL InlineProgressBarCtrl_SetRange(HWND hwnd, const long nLower, const long nUpper, const long nStep);
+int InlineProgressBarCtrl_SetStep(HWND hwnd, const long nStep);
 void InlineProgressBarCtrl_StepIt(HWND hwnd);
-void InlineProgressBarCtrl_SetPos(HWND hwnd, const int nValue);
+void InlineProgressBarCtrl_SetPos(HWND hwnd, const long nValue);
 BOOL InlineProgressBarCtrl_Resize(HWND hwnd);
